refactor : 전반적인 접근 지정자 변경 및 패키지 분리
