--- src/main/java/lotto/domain/InvalidLottoNumberException.java
@@ -1,10 +1,0 @@
-package lotto.domain;
-
-public class InvalidLottoNumberException extends IllegalArgumentException {
-	public static final String OUT_OF_BOUND = "1이상 45이하의 정수를 입력해주세요.";
-	public static final String NOT_INTEGER = "정수를 입력해주세요.";
-
-	public InvalidLottoNumberException(String s) {
-		super(s);
-	}
-}
