+++ src/main/java/lotto/domain/Lotto/LottoGenerator.java
@@ -1,0 +1,20 @@
+package lotto.domain.Lotto;
+
+import lotto.domain.LottoNumber.LottoNumber;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import java.util.stream.Collectors;
+
+public class LottoGenerator {
+    public static Lotto generate() {
+        List<LottoNumber> lottoNumbers = new ArrayList<>(LottoNumber.values());
+        Collections.shuffle(lottoNumbers);
+
+        return lottoNumbers.stream()
+                .limit(Lotto.CORRECT_SIZE)
+                .sorted()
+                .collect(Collectors.collectingAndThen(Collectors.toList(), Lotto::new));
+    }
+}
